requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
isodate==0.6.1
lxml==4.9.1
numpy==1.23.4
onvif-zeep==0.2.12
opencv-python==4.6.0.66
Pillow==9.3.0
platformdirs==2.5.4
pytz==2022.6
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
suds==1.1.2
suds-py3==1.4.5.0
tk==0.1.0
urllib3==1.26.12
zeep==4.2.0
